requirements.txt
allure-pytest==2.7.0
allure-python-commons==2.7.0
appnope==0.1.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
browsermob-proxy==0.8.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
idna==2.8
importlib-metadata==0.18
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.19
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.0.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
packaging==19.0
pandocfilters==1.4.2
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.12.0
prometheus-client==0.5.0
prompt-toolkit==2.0.8
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.8.0
Pygments==2.3.1
pylint==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.0
pytest==5.0.0
pytest-html==1.20.0
pytest-json==0.4.0
pytest-metadata==1.8.0
python-dateutil==2.8.0
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.11.1
zipp==0.5.1
